% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_mle.R
\name{ck_mle_lognormal}
\alias{ck_mle_lognormal}
\title{Lognormal MLE estimator in Czado and van Keilegom (2021) with bootstrap inference}
\usage{
ck_mle_lognormal(
  tau,
  d,
  rho,
  copula = "joe",
  x0 = NULL,
  lb = NULL,
  ub = NULL,
  xtol_rel = 1e-08,
  maxeval = 20000,
  quant_point = c(0.75, 0.5, 0.25),
  num_b = 10000,
  seed = NULL
)
}
\description{
Lognormal MLE estimator in Czado and van Keilegom (2021) with bootstrap inference
}
