% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_mle.R
\name{ck_mle_lognormal_est}
\alias{ck_mle_lognormal_est}
\title{MLE Estimator in Czado Keilegom (2021) with log-normal distributions}
\usage{
ck_mle_lognormal_est(
  tau,
  d,
  rho,
  copula = "joe",
  x0 = NULL,
  lb = NULL,
  ub = NULL,
  xtol_rel = 1e-10,
  maxeval = 20000,
  quant_point = c(0.75, 0.5, 0.25)
)
}
\arguments{
\item{tau}{min(T, C, Z)}

\item{d}{I(T < C); it is observed only when rho = 1}

\item{rho}{indicator for not censored by Z}

\item{copula}{c("joe", "clayton")}

\item{xtol_rel}{parameter for nloptr sovler}

\item{maxeval}{parameter for nloptr sovler}
}
\description{
MLE Estimator in Czado Keilegom (2021) with log-normal distributions
}
