% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_mle.R
\name{ck_mle_weibull_bootstrap}
\alias{ck_mle_weibull_bootstrap}
\title{Bootstrap Weibull MLE in Czado and van Keilegom (2021)}
\usage{
ck_mle_weibull_bootstrap(
  tau,
  d,
  rho,
  copula = "joe",
  x0 = NULL,
  lb = NULL,
  ub = NULL,
  xtol_rel = 1e-08,
  maxeval = 20000,
  quant_point = c(0.75, 0.5, 0.25),
  num_b = 10000,
  seed = NULL
)
}
\description{
Bootstrap Weibull MLE in Czado and van Keilegom (2021)
}
