% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{clayton_k_tau_to_alpha}
\alias{clayton_k_tau_to_alpha}
\title{Convert Kendall's tau to the dependence parameter alpha (Clayton)}
\usage{
clayton_k_tau_to_alpha(k_tau)
}
\description{
Convert Kendall's tau to the dependence parameter alpha (Clayton)
}
