% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{clayton_phi_deriv}
\alias{clayton_phi_deriv}
\title{First order derivative of clayton copula generator}
\usage{
clayton_phi_deriv(u, a)
}
\description{
First order derivative of clayton copula generator
}
