% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alpha.R
\name{eval_moment_log}
\alias{eval_moment_log}
\title{Compute the moments of log variables based on the pseudo model}
\usage{
eval_moment_log(theta, gamma, S, moment_sel = NULL, seed = NULL)
}
\arguments{
\item{theta}{parameters of the log-normal distributions of T and C}

\item{gamma}{parameter of the exponential distribution of P}

\item{S}{number of sample draws}

\item{moment_sel}{index of moments in use}
}
\description{
Compute the moments of log variables based on the pseudo model
}
