% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alpha.R
\name{generate_tc_lognormal}
\alias{generate_tc_lognormal}
\title{Generate T and C based log-normal parameters and p}
\usage{
generate_tc_lognormal(theta, p, min_sig = 1e-04)
}
\arguments{
\item{p}{}
}
\description{
Generate T and C based log-normal parameters and p
}
