% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{inverse_cdf}
\alias{inverse_cdf}
\title{generate data by inverse a estimated CDF function}
\usage{
inverse_cdf(fi, xi, n, seed = NULL, round_digits = 5)
}
\description{
generate data by inverse a estimated CDF function
}
