% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{joe_alpha_to_k_tau}
\alias{joe_alpha_to_k_tau}
\title{Convert dependence parameter alpha (Joe) to Kendall's tau}
\usage{
joe_alpha_to_k_tau(alpha)
}
\description{
Convert dependence parameter alpha (Joe) to Kendall's tau
}
