% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{joe_k_tau_to_alpha}
\alias{joe_k_tau_to_alpha}
\title{Convert Kendall's tau to the dependence parameter alpha (Joe)}
\usage{
joe_k_tau_to_alpha(k_tau)
}
\description{
Convert Kendall's tau to the dependence parameter alpha (Joe)
}
