% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{km}
\alias{km}
\title{Bootstrapped Kaplan-Meier}
\usage{
km(
  tau,
  d,
  sig_level = 0.05,
  quant_point = c(0.25, 0.5, 0.75),
  num_b = 10000,
  seed = NULL
)
}
\arguments{
\item{tau}{min(T, C)}

\item{d}{I(T < C)}
}
\value{
A list
\item{quant_time}{Estimated quantiles of survival times}
\item{quant_time_bt}{Bootsrapped estimated quantiles}
\item{CI}{Confidence Intervals}
}
\description{
Bootstrapped Kaplan-Meier
}
