% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{km_est}
\alias{km_est}
\title{Kaplan-Meier Estimation}
\usage{
km_est(tau, d)
}
\arguments{
\item{tau}{min(T, C)}

\item{d}{I(T < C)}
}
\value{
survival probability and survival time for those with d == 1 with
repeated entries removed
}
\description{
Mimicing Matlab ecdf function.
}
