% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_mle.R
\name{lognormal_mle}
\alias{lognormal_mle}
\title{lognormal MLE}
\usage{
lognormal_mle(
  x,
  x0 = c(0, 1),
  lb = c(-10, 0),
  ub = c(10, 5),
  xtol_rel = 1e-10,
  maxeval = 20000
)
}
\description{
lognormal MLE
}
