% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alpha.R
\name{md_est}
\alias{md_est}
\title{Minimum Distance Estimation}
\usage{
md_est(
  m_sample,
  S,
  moment_sel,
  lb,
  ub,
  para_init,
  range_factor,
  seed,
  num_rand_perturb,
  x_tol_rel = 1e-08,
  maxeval = 10000
)
}
\description{
Minimum Distance Estimation
}
