% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alpha.R
\name{min_dist_obj}
\alias{min_dist_obj}
\title{Objective function of minimum distance (Log normal T and C and log moments)}
\usage{
min_dist_obj(para, S, m_sample, moment_sel)
}
\description{
Objective function of minimum distance (Log normal T and C and log moments)
}
