% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alpha.R
\name{mm_alpha}
\alias{mm_alpha}
\title{Pseudo Model Method of Moments to choose the depedent parameter alpha}
\usage{
mm_alpha(
  tau,
  d,
  p,
  moment_options_vec = 1:6,
  S = 20000,
  lb = c(-10, 0, 0, -10, 0, 0, 0),
  ub = c(10, 5, 5, 10, 5, 5, 1),
  para_init = c(0.5, 1, 0.4, 0.5, 1, 0.4, 0.0055),
  range_factor = c(10, 2, 2, 10, 2, 2, 500),
  seed = 99,
  num_rand_perturb = 100,
  two_stage = FALSE,
  min_sig = 0.25,
  x_tol_rel = 1e-08,
  maxeval = 10000
)
}
\arguments{
\item{tau}{min{T, C} without random censoring and Delta_P = 1}

\item{d}{I(T < C)}

\item{p}{disease progress}

\item{S}{= 20000}

\item{maxeval}{}
}
\description{
Pseudo Model Method of Moments to choose the depedent parameter alpha
}
