% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quantile_surv_time}
\alias{quantile_surv_time}
\title{Compute the quantiles of survival time}
\usage{
quantile_surv_time(surv, time, quant_point = c(0.25, 0.5, 0.75))
}
\arguments{
\item{quant_point}{}
}
\description{
Caveat: Note that the estimated survival function and time are sorted
}
