% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_graphic.R
\name{rw_bootstrap}
\alias{rw_bootstrap}
\title{Bootstrap Copula Graphic Estimator}
\usage{
rw_bootstrap(
  tau,
  d,
  rho,
  alpha,
  copula = "joe",
  t_eval = NULL,
  quant_point = c(0.25, 0.5, 0.75),
  num_b = 10000,
  seed = NULL
)
}
\arguments{
\item{tau}{min(T, C, Z)}

\item{d}{I(T < C); it is observed only when rho = 1}

\item{rho}{indicator for not censored by Z}

\item{alpha}{Joe Copula parameter (scalar >= 1)}

\item{copula}{c("joe", "clayton")}

\item{num_b}{number of Bootstrap sampling}

\item{seed}{}
}
\value{
A list
\item{quantile}{A matrix of bootstrapped quantiles}
\item{surv_eval}{A matrix of bootstrapped survival fucntion evaluated at t_eval}
}
\description{
Bootstrap Copula Graphic Estimator
}
