% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_graphic.R
\name{rw_est}
\alias{rw_est}
\title{Copula Graphic Estimator}
\usage{
rw_est(tau, d, rho, alpha, copula = "joe")
}
\arguments{
\item{tau}{min(T, C, Z)}

\item{d}{I(T < C); it is observed only when rho = 1}

\item{rho}{indicator for not censored by Z}

\item{alpha}{Joe Copula parameter (scalar >= 1)}

\item{copula}{c("joe", "clayton")}
}
\value{
A data.frame
\item{surv}{n_1-by-1 vector estimated survival function}
\item{time}{survival time}
}
\description{
Modified Rivest and Wells (2001)'s estimator
using data (t_i, d_i) of individuals with rho_i = 1 to estimate the
survival function of the event at points t_(1), t_(2), ..., t_(n_1)
}
