% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alpha.R
\name{sample_moments_log}
\alias{sample_moments_log}
\title{Compute the sample moments of log variables}
\usage{
sample_moments_log(tau, d, p, moment_sel = NULL)
}
\arguments{
\item{tau}{min{T, C} without random censoring and Delta_P = 1}

\item{d}{I(T < C)}

\item{p}{disease progress}
}
\description{
Compute the sample moments of log variables
}
