% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{surv_func_eval}
\alias{surv_func_eval}
\title{Evaluate Survival Function}
\usage{
surv_func_eval(surv_est, time_est, time_eval)
}
\arguments{
\item{time_eval}{}
}
\description{
Evaluate Survival Function
}
