% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{surv_plot}
\alias{surv_plot}
\title{Plot the estimated survival function}
\usage{
surv_plot(time, surv, x_interval = 100)
}
\arguments{
\item{x_interval}{Interval length on x-axis}
}
\description{
Plot the estimated survival function
}
