% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_mle.R
\name{weibull_mle}
\alias{weibull_mle}
\title{Weibull MLE}
\usage{
weibull_mle(
  x,
  x0 = c(1, mean(x)),
  lb = c(0, 0),
  ub = c(10, max(x)),
  xtol_rel = 1e-08,
  maxeval = 20000
)
}
\description{
Weibull MLE
}
